requirements.txt
torch==2.2.1
torchvision==0.17.1
einops==0.7.0
decord==0.6.0
diffusers==0.27.2
packaging==24.0
omegaconf==2.3.0
transformers==4.39.0
accelerate==0.28.0
pandas==2.2.1
wandb==0.16.4
opencv-python==4.9.0.80
moviepy==1.0.3
imageio==2.34.0
scipy==1.12.0
scikit-learn==1.4.1.post1
matplotlib-3.8.3